Create README.md
Sentiment Analysis is an important algorithm in Natural Language Processing that is used to detect sentiment within some text. In our project, we had chosen to work on analyzing reviews of various drugs which have been reviewed in form of texts and have also been given a rating on a scale from 1-10.